services/dashboard/requirements.txt
# services/dashboard/requirements.txt
# HoneyChat 仪表盘服务依赖清单

# Web 框架
Flask==3.0.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
python-socketio==5.11.1
python-engineio==4.8.0

# 数据库
SQLAlchemy==2.0.27
alembic==1.13.1  # 用于数据库迁移（可选）

# WebSocket 服务器（生产环境推荐使用 eventlet 或 gevent）
eventlet==0.35.1
# gevent==24.2.1  # 也可选 gevent，但需与 gunicorn 配合

# WSGI 服务器
gunicorn==21.2.0

# 配置管理
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
PyYAML==6.0.1

# 日志处理
structlog==24.1.0
python-json-logger==2.0.7

# 健康检查依赖
requests==2.31.0

# 类型提示
typing-extensions==4.10.0

# 工具库
python-dateutil==2.8.2